Refactor openapi 3 helpers into OpenApi3Specification class
