+++ spec/support/vendor_api/vendor_api_spec_helpers.rb
@@ -7,65 +7,23 @@ module VendorApiSpecHelpers
     parsed_response['errors'].first
   end
 
-  def load_openapi_spec(path)
-    YAML.load_file(path)
-  end
-
-  def transform_openapi_schema_to_json_schema(schema)
-    properties = schema['properties']
-    schema['properties'] = properties.reduce({}) do |new_props, (prop, value)|
-      new_props[prop] = if value['nullable'] == 'true'
-                          {
-                            'oneOf' => [
-                              value.except('nullable'),
-                              { 'type' => 'null' },
-                            ],
-                          }
-                        else
-                          value
-                        end
-      new_props
-    end
-
-    schema
-  end
-
-  def parse_openapi_json_schema(spec, schema_name)
-    # Pull up the schema that we want to validate against into the top-level,
-    # so that json-schema understands it.
-    spec['$schema'] = 'http://json-schema.org/draft-04/schema#'
-    spec['$ref'] = "#/components/schemas/#{schema_name}"
-
-    schemas = spec['components']['schemas']
-    transformed_schemas = schemas.reduce({}) do |new_schemas, (name, definition)|
-      new_schemas[name] = transform_openapi_schema_to_json_schema(definition)
-      new_schemas
-    end
-
-    spec['components']['schemas'] = transformed_schemas
-    spec
-  end
-
   RSpec::Matchers.define :be_valid_against_openapi_schema do |schema_name|
     match do |item|
-      schema = parse_openapi_json_schema(
-        load_openapi_spec("#{Rails.root}/config/vendor-api-0.4.0.yml"),
-        schema_name,
-      )
+      spec = OpenApi3Specification.new(YAML.load_file("#{Rails.root}/config/vendor-api-0.4.0.yml"))
 
       JSONSchemaValidator.new(
-        schema,
+        spec.as_json_schema(schema_name),
         item,
       ).valid?
     end
 
     failure_message do |item|
-      schema = parse_openapi_json_schema(
-        load_openapi_spec("#{Rails.root}/config/vendor-api-0.4.0.yml"),
-        schema_name,
-      )
+      spec = OpenApi3Specification.new(YAML.load_file("#{Rails.root}/config/vendor-api-0.4.0.yml"))
 
-      JSONSchemaValidator.new(schema, item).failure_message
+      JSONSchemaValidator.new(
+        spec.as_json_schema(schema_name),
+        item
+      ).failure_message
     end
   end
 
@@ -82,7 +40,6 @@ module VendorApiSpecHelpers
     end
 
     def failure_message
-      # @TODO fix this error message so it names the schema
       <<~ERROR
         Expected the item to be valid against schema:
